Rewrite rewards_report from set returning function to cursor based stored proc.
The original rewards report was a stored procedure that returned multiple
result sets. Since postgres lacks this capability, at the time I rewrote the
function to just return a single result of the customer details. For PG11, I
wanted to rewrite this as an official stored procedure, however stored procs
cannot return full result sets, so I rewrote it to return multiple cursor
references instead. This requires some semantic changes as we now must run the
procedure within a transaction block, but it also shows off some of the
capabilities of the system we haven't used before. A proper example of use
would be:

    begin;
      call rewards_report(5,25,'2007-01-01'::date);
      fetch all in rewardees_count;
      fetch all in rewardees_detail;
    commit;

To be fair, the multiple refcursor approach could be implemented within a
Postgres function, however this is perhaps the only example of this usage in a
stored procedure, so I think the example is still valuable.

Note I also fixed some logic bugs and left some debugging examples within the
new code as well.