Initial clean dump with pg_dump version 12.
The primary result is that trigger definitions now use 'EXECUTE FUNCTION'
rather than the previous syntax of 'EXECUTE PROCEDURE'. This is a left-over
nod to the implementation of "true" stored procedures in Postgres 11. It may
be worth mentioning that Postgres 12 does not add the ability to call stored
procedures from a trigger, so this is essentially just a cosmetic change at
this point.