convert customer.active to generated column
In theory the active and activebool columns should always show the same value,
but we've never enforced that. Here we convert the the actice column to
smallint and make it reflect what we are doing in the activebool version of
the column. We also updated the data to no longer include the active column,
as that will come from the generated column command. As an added bonus, this
gives us a non-dientity example of using generated columns.