utility sproc to set payments to recent date range
While I have traditionaly kept the same date range for continuity, it is
easier for most folks to work with current/recent dates, especially in
conjunction with now() and similar functions, so this sproc does a quick and
direty job of updating all the payments to recent dates.

I had a few different versions of this over time, and while there might be
interesting things to glean from the procedure code itself, the code here is
pretty grotty, so I wouldn't really recommend it for that.